Move some effect hooks up
